refactor: 결제 검증 기능 수정 (#8)

+++ src/main/java/com/example/demo/src/payment/entity/enums/model/PostPayReq.java
@@ -1,8 +1,8 @@
 package com.example.demo.src.payment.entity.enums.model;
 
+import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -13,5 +13,5 @@ public class PostPayReq {
 
     private String impUid;
     private String merchantUid;
-    private BigDecimal amount;
+    private int amount;
 }
